// src/generated/models/Intake_perfilsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Intake_perfilsstatecode = {
  0: 'Ativo_a_',
  1: 'Inativo_a_'
} as const;
export type Intake_perfilsstatecode = keyof typeof Intake_perfilsstatecode;
export const Intake_perfilsstatuscode = {
  1: 'Ativo_a_',
  2: 'Inativo_a_'
} as const;
export type Intake_perfilsstatuscode = keyof typeof Intake_perfilsstatuscode;

export interface Intake_perfilsBase {
  importsequencenumber?: number;
  intake_ativo?: boolean;
  intake_frentesjson?: string;
  intake_nome?: string;
  intake_papeisjson?: string;
  intake_perfilid: string;
  intake_upn: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Intake_perfilsstatecode;
  statuscode?: Intake_perfilsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Intake_perfils extends Intake_perfilsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  intake_ativoname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}